refactor(agent): split session event dispatcher
